+++ lib/td_bg/business_concepts/business_concepts.ex
@@ -861,14 +861,17 @@ defmodule TdBg.BusinessConcepts do
     |> Repo.transaction()
   end
 
-  def get_business_concept_by_name(name) do
-    BusinessConceptVersion
-    |> join(:left, [v], _ in assoc(v, :business_concept))
-    |> join(:left, [v, c], _ in assoc(c, :domain))
-    |> where([v], ilike(v.name, ^"%#{name}%"))
-    |> preload([_, c, d], business_concept: {c, domain: d})
-    |> order_by(asc: :version)
-    |> Repo.all()
+  def get_concept_by_name_in_domain(name, domain_id) do
+    BusinessConcept
+    |> join(:inner, [c], v in assoc(c, :versions))
+    |> join(:left, [c, v], d in assoc(c, :domain))
+    |> where([c, v], v.name == ^to_string(name) and c.domain_id == ^domain_id)
+    |> preload([c, _v, d],
+      domain: d,
+      shared_to: [],
+      versions: ^from(v in BusinessConceptVersion, order_by: [desc: v.version])
+    )
+    |> Repo.one()
   end
 
   def get_template(%BusinessConceptVersion{business_concept: business_concept}) do